De base n a base 10 terminado

+++ De_Base_N_a_10.h
@@ -3,5 +3,5 @@ void deNa10Decimal(int *BaseOrigen,char *numeroDecimal, char *resultadoNa10Strin
 void deNa10Entero(int *BaseOrigen,char *numeroEntero, char *stringResultadoNa10);
 void concatenarCadenas(char *numEntero, char *numDecimal, char *numCompleto);
 void mapearDigito(char * pDigito, int *digitoInt);
-void toString(float *resultadoFloat, char * resultadoString, char * parteNum );
+void toString(double *resultadoFloat, char * resultadoString, char * parteNum );
 void dividirCadenas(char *numEntero, char *numDecimal, char *numCompleto);
